+++ src/Api/TradeDataApi/TradeImportsDataApiOptions.cs
@@ -1,6 +1,6 @@
 namespace Defra.PhaImportNotifications.Api.TradeDataApi;
 
-public class TradeDataApiOptions
+public class TradeImportsDataApiOptions
 {
     public required string BaseUrl { get; set; }
     public string? Username { get; set; }
